TestHomeViewをtweets/test.pyに記述
