+++ accounts/tests.py
@@ -247,25 +247,3 @@ class TestSignupView(TestCase):
         self.assertFalse(form.is_valid())
         self.assertIn("確認用パスワードが一致しません。", form.errors["password2"])
         # 「確認用パスワードが一致しません。」などのエラー文は少しでも間違えるとテストでエラーが出るので気を付けること
-
-
-class TestHomeView(TestCase):
-
-    def setUp(self):
-        self.url = reverse("tweets:home")
-
-        # ユーザーを作成
-        self.user = User.objects.create_user(username="tester", password="testpassword")
-
-        # そのユーザーをログイン
-        self.client.login(username="tester", password="testpassword")
-
-    # リクエストの送信
-    def test_success_get(self):
-
-        response = self.client.get(self.url)
-        # self.url = tweets/home
-        # self.client = サーバーに対してリクエストをシミュレート
-
-        self.assertEqual(response.status_code, 200)
-        # assertEqual(A, B) = Aの結果とBが一致すれば成功
