+++ tweets/tests.py
@@ -1,44 +1,27 @@
-# from django.test import TestCase
+from django.contrib.auth import get_user_model
+from django.test import TestCase
+from django.urls import reverse
+
+User = get_user_model()
 
 
-# class TestHomeView(TestCase):
-#     def test_success_get(self):
+class TestHomeView(TestCase):
 
+    def setUp(self):
+        self.url = reverse("tweets:home")
 
-# class TestTweetCreateView(TestCase):
-#     def test_success_get(self):
+        # ユーザーを作成
+        self.user = User.objects.create_user(username="tester", password="testpassword")
 
-#     def test_success_post(self):
+        # そのユーザーをログイン
+        self.client.login(username="tester", password="testpassword")
 
-#     def test_failure_post_with_empty_content(self):
+    # リクエストの送信
+    def test_success_get(self):
 
-#     def test_failure_post_with_too_long_content(self):
+        response = self.client.get(self.url)
+        # self.url = tweets/home
+        # self.client = サーバーに対してリクエストをシミュレート
 
-
-# class TestTweetDetailView(TestCase):
-#     def test_success_get(self):
-
-
-# class TestTweetDeleteView(TestCase):
-#     def test_success_post(self):
-
-#     def test_failure_post_with_not_exist_tweet(self):
-
-#     def test_failure_post_with_incorrect_user(self):
-
-
-# class TestLikeView(TestCase):
-#     def test_success_post(self):
-
-#     def test_failure_post_with_not_exist_tweet(self):
-
-#     def test_failure_post_with_liked_tweet(self):
-
-
-# class TestUnLikeView(TestCase):
-
-#     def test_success_post(self):
-
-#     def test_failure_post_with_not_exist_tweet(self):
-
-#     def test_failure_post_with_unliked_tweet(self):
+        self.assertEqual(response.status_code, 200)
+        # assertEqual(A, B) = Aの結果とBが一致すれば成功
